backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Limiter==3.3.0
Flask-Migrate==4.0.4
python-dotenv==1.0.0
PyJWT==2.7.0
itsdangerous==2.1.2
numpy==1.24.3
Pillow==9.5.0
opencv-python-headless==4.7.0.72
redis==4.5.5
gunicorn==20.1.0
psycopg2-binary==2.9.6
pydantic==1.10.7